Adding new setup_apt.sh for redhat

+++ platform/tools/drivers-linux-eth/setup_apt.sh
@@ -1,5 +1,30 @@
 #!/bin/bash
-apt-get install -y \
+
+ exists()
+{
+  command -v "$1" >/dev/null 2>&1
+}
+
+ if exists yum; then
+  yum install -y \
+    libtool \
+    automake \
+    autoconf \
+    cmake \
+    gcc \
+    libnl3-devel \
+    libudev-devel \
+    make \
+    pkgconfig \
+    valgrind-devel \
+    iperf3 \
+    vim \
+    sshpass \
+    sysfsutils \
+    net-tools
+
+ else
+   apt-get install -y \
     build-essential \
     libtool \
     automake \
@@ -12,3 +37,4 @@ apt-get install -y \
     ninja-build \
     pkg-config \
     valgrind
+fi
